all doc except examples :)
This one will be pretty quick, getting pretty close to completing it. Then, soon, I will have `Gattino`. Also currently doing `ToolipsSession`, which will be the final `Toolips`-side step towards building the new `Olive` in `0.3` (yes, seriously, that's it! -- though I am also planning to take some time with `IPyCells` and the highlighters as well -- (which of course are being moved)